requirements.txt
Django>=5.0,<6.0
wagtail>=7.1,<8.0
psycopg[binary,pool]>=3.1
django-redis>=5.4
djangorestframework>=3.15
django-taggit>=4.0
django-storages[boto3]>=1.14
boto3>=1.34
celery>=5.3
redis>=5.0
opensearch-py>=2.4
clickhouse-driver>=0.2.9
python-dotenv>=1.0
gunicorn>=22.0
django-cors-headers>=4.7
PyYAML>=6.0
jieba>=0.42.1
PyJWT>=2.8.0
